+++ src/weapon/construct.h
@@ -31,7 +31,7 @@ class Construct : public Weapon
 private:
   bool target_chosen;
   Sprite* construct_spr;
-  int angle;
+  double angle;
   Point2i dst;
 
 protected:
